+++ pkg/dataplane/identity_test.go
@@ -1,0 +1,64 @@
+package dataplane
+
+import (
+	"errors"
+	"testing"
+
+	"github.com/Azure/msi-dataplane/internal/swagger"
+	"github.com/Azure/msi-dataplane/internal/test"
+)
+
+func TestValidateUserAssignedMSI(t *testing.T) {
+	t.Parallel()
+
+	testCases := []struct {
+		name        string
+		getMSI      func() []*swagger.NestedCredentialsObject
+		resourceIDs []string
+		expectedErr error
+	}{
+		{
+			name:        "nil credential object slice",
+			getMSI:      func() []*swagger.NestedCredentialsObject { return nil },
+			resourceIDs: []string{"someResourceID"},
+			expectedErr: errNumberOfMSIs,
+		},
+		{
+			name: "nil fields",
+			getMSI: func() []*swagger.NestedCredentialsObject {
+				return []*swagger.NestedCredentialsObject{&swagger.NestedCredentialsObject{}}
+			},
+			resourceIDs: []string{"someResourceID"},
+			expectedErr: errNilField,
+		},
+		{
+			name: "mismatched resourceID",
+			getMSI: func() []*swagger.NestedCredentialsObject {
+				testMSI := test.GetTestMSI("bogus")
+				return []*swagger.NestedCredentialsObject{testMSI}
+			},
+			resourceIDs: []string{"someResourceID"},
+			expectedErr: errResourceIDNotFound,
+		},
+		{
+			name: "success",
+			getMSI: func() []*swagger.NestedCredentialsObject {
+				testMSI := test.GetTestMSI(validResourceID)
+				return []*swagger.NestedCredentialsObject{testMSI}
+			},
+			resourceIDs: []string{validResourceID},
+			expectedErr: nil,
+		},
+	}
+
+	for _, tc := range testCases {
+		tc := tc
+		t.Run(tc.name, func(t *testing.T) {
+			t.Parallel()
+			msi := tc.getMSI()
+			if err := validateUserAssignedMSIs(msi, tc.resourceIDs); !errors.Is(err, tc.expectedErr) {
+				t.Errorf("expected error: `%s` but got: `%s`", tc.expectedErr, err)
+			}
+		})
+	}
+}
